+++ packages/runtime/src/widgets/Avatar/Avatar.tsx
@@ -1,19 +1,16 @@
-import React, { useEffect, useState } from "react";
+import React, { useCallback, useMemo, useState } from "react";
 import type { EnsembleAction, Expression } from "@ensembleui/react-framework";
-import { useRegisterBindings } from "@ensembleui/react-framework";
+import { isExpression, useRegisterBindings } from "@ensembleui/react-framework";
 import {
   Avatar as MuiAvatar,
   Menu,
   MenuItem,
   ListItemIcon,
 } from "@mui/material";
-import { isString } from "lodash-es";
-import { useNavigateScreen } from "../../runtime/hooks/useNavigateScreen";
 import { WidgetRegistry } from "../../registry";
 import type { EnsembleWidgetStyles, IconProps } from "../../shared/types";
 import { Icon } from "../Icon";
-import { useExecuteCode } from "../../runtime/hooks/useEnsembleAction";
-import { stringToColor } from "./utils/stringToColors";
+import { useEnsembleAction } from "../../runtime/hooks/useEnsembleAction";
 import { generateInitials } from "./utils/generateInitials";
 
 const widgetName = "Avatar";
@@ -36,8 +33,6 @@ export interface AvatarProps {
 }
 
 export const Avatar: React.FC<AvatarProps> = (props) => {
-  const [code, setCode] = useState("");
-  const [screen, setScreen] = useState("");
   const [menuAnchorEl, setMenuAnchorEl] = useState<null | HTMLElement>(null);
   const [isMenuOpen, setIsMenuOpen] = useState<boolean>(Boolean(menuAnchorEl));
   const [name, setName] = useState(props.name);
@@ -48,9 +43,11 @@ export const Avatar: React.FC<AvatarProps> = (props) => {
       setName,
     },
   );
-  // FIXME: action callbacks should take params so they can be callable per element
-  const executeCode = useExecuteCode(code, { context: values });
-  const navigate = useNavigateScreen(screen);
+
+  const actions = values?.menu?.map((menuItem) => {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    return useEnsembleAction(menuItem.onTap);
+  });
 
   const handleMenuOpen = (event: React.MouseEvent<HTMLElement>): void => {
     setMenuAnchorEl(event.currentTarget);
@@ -62,65 +59,57 @@ export const Avatar: React.FC<AvatarProps> = (props) => {
     setIsMenuOpen(false);
   };
 
-  const handleMenuClick = (menuItem: AvatarMenu): void => {
-    menuItem.onTap &&
-      "executeCode" in menuItem.onTap &&
-      setCode(String(menuItem.onTap.executeCode));
-    menuItem.onTap &&
-      "navigateScreen" in menuItem.onTap &&
-      isString(menuItem.onTap.navigateScreen) &&
-      setScreen(menuItem.onTap.navigateScreen);
-    handleMenuClose();
-  };
+  const handleMenuClick = useCallback(
+    (menuItem: AvatarMenu): void => {
+      const menuIndex = values?.menu?.indexOf(menuItem);
+      if (menuIndex !== undefined) {
+        actions?.at(menuIndex)?.callback({
+          ...menuItem,
+        });
+      }
 
-  useEffect(() => {
-    code && executeCode?.callback();
-  }, [code, executeCode]);
-  useEffect(() => {
-    screen && navigate?.callback();
-  }, [screen, navigate]);
+      handleMenuClose();
+    },
+    [values?.menu, actions],
+  );
+
+  const initials = useMemo(
+    () => (isExpression(values?.name) ? "" : generateInitials(values?.name)),
+    [values?.name],
+  );
 
   return (
     <div ref={rootRef}>
       <MuiAvatar
-        alt={props.alt}
+        alt={values?.alt}
         onClick={handleMenuOpen}
-        src={props.src}
+        src={values?.src}
         sx={{
-          bgcolor:
-            props.styles?.backgroundColor ?? stringToColor(values?.name ?? ""),
-          width: props.styles?.width,
-          height: props.styles?.height,
-          color: props.styles?.color ?? "white",
+          backgroundColor: "black",
+          color: "white",
           cursor: "pointer",
-          ...values?.styles,
           ...values?.styles,
           ...(values?.styles?.visible === false
             ? { display: "none" }
             : undefined),
         }}
       >
-        {props.name ? (
-          generateInitials(values?.name)
-        ) : (
+        {initials || (
           <Icon
-            color={props.icon?.color}
-            name={props.icon?.name ?? ""}
-            size={props.icon?.size}
+            color={values?.icon?.color}
+            name={values?.icon?.name ?? ""}
+            size={values?.icon?.size}
           />
         )}
       </MuiAvatar>
-      {props.menu ? (
+      {values?.menu ? (
         <Menu
           anchorEl={menuAnchorEl}
           onClose={handleMenuClose}
           open={isMenuOpen}
         >
-          {props.menu.map((menuItem) => (
-            <MenuItem
-              key={menuItem.label}
-              onClick={(): void => handleMenuClick(menuItem)}
-            >
+          {values?.menu.map((menuItem, idx) => (
+            <MenuItem key={idx} onClick={(): void => handleMenuClick(menuItem)}>
               {menuItem.icon ? (
                 <ListItemIcon>
                   <Icon